requirements.txt
blinker==1.9.0
click==8.2.1
ecdsa==0.19.1
Flask==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.4.0
packaging==25.0
python-dotenv==1.1.1
python-http-client==3.3.7
sendgrid==6.12.4
six==1.17.0
Werkzeug==3.1.3
